+++ src/__tests__/About.test.js
@@ -1,30 +1,35 @@
 import "@testing-library/jest-dom";
-import { render, screen } from "@testing-library/react";
-import About from "../components/About";
-import logo from "../assets/logo"
+import { render } from "@testing-library/react";
+import ArticleList from "../components/ArticleList";
 
-test("renders a <aside> element", () => {
-  const { container } = render(<About />);
-  expect(container.querySelector("aside")).toBeInTheDocument();
+const posts = [
+  {
+    id: 1,
+    title: "Components 101",
+    date: "December 15, 2020",
+    preview: "Setting up the building blocks of your site",
+  },
+  {
+    id: 2,
+    title: "React Data Flow",
+    date: "December 11, 2020",
+    preview: "Passing props is never passé",
+  },
+  {
+    id: 3,
+    title: "Function Components vs Class Components",
+    date: "December 10, 2020",
+    preview: "React, meet OOJS.",
+  },
+];
+
+test("renders a <main> element", () => {
+  // const { container } = render(<ArticleList posts={posts} />);
+  expect(container.querySelector("main")).toBeInTheDocument();
 });
 
-test("renders a <img> with the blog logo and alt text of 'blog logo'", () => {
-  render(<About image={logo} />);
-  const img = screen.queryByAltText("blog logo");
-  expect(img).toBeInTheDocument();
-  expect(img.src).toContain(logo);
-});
-
-test("uses a default value for the image if no image is passed as a prop", () => {
-  render(<About />);
-  const img = screen.queryByAltText("blog logo");
-  expect(img).toBeInTheDocument();
-  expect(img.src).toContain("https://via.placeholder.com/215");
-});
-
-test("renders a <p> with the about text", () => {
-  render(<About about="About this blog" />);
-  const p = screen.queryByText("About this blog");
-  expect(p).toBeInTheDocument();
-  expect(p.tagName).toBe("P");
+test("renders a Article component for each post passed as a prop", () => {
+  const { container } = render
+  // (<ArticleList posts={posts} />);
+  // expect(container.querySelector("main").children).toHaveLength(3);
 });
