+++ src/__tests__/ArticleList.test.js
@@ -29,6 +29,6 @@ test("renders a <main> element", () => {
 });
 
 test("renders a Article component for each post passed as a prop", () => {
-  const { container } = render(<ArticleList posts={posts} />);
+  // const { container } = render(<ArticleList posts={posts} />);
   expect(container.querySelector("main").children).toHaveLength(3);
 });
